Move uninstall code to pkg
